build: :heavy_plus_sign: Adding testing dependencies, end a new script
adding necessaries depedencies to code works

Add testing-library/react, testing-library/react-hooks, babel-jest, and a script to install all dependencies  Created this script to install all dependencies, as some of them are not up to date and are only needed for testing purposes. The dependencies in question are:  testing-library/react testing-library/react-hooks According to their developers, they are already in the process of being updated.